created a new job role detail response and change open job response to parent job role response, then inherited it
