+++ src/main/java/org/example/models/JobRoleResponse.java
@@ -3,6 +3,7 @@ package org.example.models;
 import java.sql.Date;
 
 public class JobRoleResponse {
+    private int jobRoleID;
     /*
     instantiates a string of Job Role Name in the Job Role Response
      */
@@ -20,40 +21,32 @@ public class JobRoleResponse {
      */
     private String jobRoleBand;
     /*
-    instantiates a Job Role CLOSING Date in the Job Role Response
+    instantiates a Date of Job Role CLOSING Date in the Job Role Response
      */
     private Date jobRoleClosingDate;
-    /*
-    instantiates a job role spec Url in the Job Role Response
-     */
-    private String jobRoleSpecUrl;
-    /*
-    instantiates a job roles responsibilities in the Job Role Response
-     */
-    private String jobRoleResponsibilities;
-    /*
-    instantiates a Date of Job Role Description Date in the Job Role Response
-     */
-    private String jobRoleDescription;
 
 
     public JobRoleResponse(
+            final int jRID,
             final String jRName,
             final Locations jRLocation,
             final String jRCapability,
             final String jRBand,
-            final Date jRClosingDate,
-            final String jRSpecUrl,
-            final String jRResponsibilities,
-            final String jRDescription) {
+            final Date jRClosingDate) {
+        this.jobRoleID = jRID;
         this.jobRoleName = jRName;
         this.jobRoleLocation = jRLocation;
         this.jobRoleCapability = jRCapability;
         this.jobRoleBand = jRBand;
         this.jobRoleClosingDate = jRClosingDate;
-        this.jobRoleSpecUrl = jRSpecUrl;
-        this.jobRoleResponsibilities = jRResponsibilities;
-        this.jobRoleDescription = jRDescription;
+    }
+
+    public int getJobRoleID() {
+        return jobRoleID;
+    }
+
+    public void setJobRoleID(final int jobRoleID) {
+        this.jobRoleID = jobRoleID;
     }
 
     public String getJobRoleName() {
@@ -94,30 +87,6 @@ public class JobRoleResponse {
 
     public void setJobRoleClosingDate(final Date jobRoleClosingDate) {
         this.jobRoleClosingDate = jobRoleClosingDate;
-    }
-
-    public String getJobRoleSpecUrl() {
-        return jobRoleSpecUrl;
-    }
-
-    public void setJobRoleSpecUrl(final String jobRoleSpecUrl) {
-        this.jobRoleSpecUrl = jobRoleSpecUrl;
-    }
-
-    public String getJobRoleResponsibilities() {
-        return jobRoleResponsibilities;
-    }
-
-    public void setJobRoleResponsibilities(final String jobRoleResponsibilities) {
-        this.jobRoleResponsibilities = jobRoleResponsibilities;
-    }
-
-    public String getJobRoleDescription() {
-        return jobRoleDescription;
-    }
-
-    public void setJobRoleDescription(final String jobRoleDescription) {
-        this.jobRoleDescription = jobRoleDescription;
     }
 }
