+++ src/main/java/org/example/services/JobRoleService.java
@@ -4,8 +4,8 @@ import org.example.Exceptions.DoesNotExistException;
 import org.example.Exceptions.Entity;
 import org.example.daos.JobRoleDao;
 import org.example.exceptions.DatabaseConnectionException;
+import org.example.models.JobRoleDetailResponse;
 import org.example.models.JobRoleResponse;
-import org.example.models.OpenJobRoleResponse;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -17,19 +17,19 @@ public class JobRoleService {
         this.jobRoleDao = jobRoleDao;
     }
 
-    public List<OpenJobRoleResponse> getOpenJobRoles()
+    public List<JobRoleResponse> getOpenJobRoles()
             throws SQLException, DatabaseConnectionException {
         return jobRoleDao.getOpenJobRoles();
     }
 
-    public JobRoleResponse getJobRoleById(final int id)
+    public JobRoleDetailResponse getJobRoleById(final int id)
             throws SQLException, DatabaseConnectionException,
             DoesNotExistException {
-        JobRoleResponse jobRoleResponse = jobRoleDao.getJobRoleByID(id);
-        if (jobRoleResponse == null) {
+        JobRoleDetailResponse jobRoleDetailResponse = jobRoleDao.getJobRoleByID(id);
+        if (jobRoleDetailResponse == null) {
             throw new DoesNotExistException(Entity.JOBROLERESPONSE);
         }
-        return jobRoleResponse;
+        return jobRoleDetailResponse;
     }
